Correction due to cherry-pick: suppress long long int and " = 0" at end line

+++ src/cpp/matrix/MatrixCOO.cpp
@@ -384,7 +384,7 @@ void tbsla::cpp::MatrixCOO::fill_cdistrib(int n_row, int n_col, int nnz, int pr,
 }
 
 
-void tbsla::cpp::MatrixCOO::fill_random_symmetric(long long int n_row, long long int n_col, long long int  nnz, long long int pr, long long int pc, long long int NR, long long int NC){
+void tbsla::cpp::MatrixCOO::fill_random_symmetric(int n_row, int n_col, int  nnz, int pr, int pc, int NR, int NC){
 }
 // TODO : normalization for COO
 void tbsla::cpp::MatrixCOO::get_row_sums(double* s) {
