+++ src/cpp/matrix/MatrixELL.cpp
@@ -879,5 +879,5 @@ void tbsla::cpp::MatrixELL::fill_cdistrib(int n_row, int n_col, int nnz, int pr,
 
 }
 
-void tbsla::cpp::MatrixELL::fill_random_symmetric(long long int n_row, long long int n_col, long long int  nnz, long long int pr, long long int pc, long long int NR, long long int NC){
+void tbsla::cpp::MatrixELL::fill_random_symmetric(int n_row, int n_col, int  nnz, int pr, int pc, int NR, int NC){
 }
